🚚 move sections & replace useless tsx by ts
