+++ packages/ui/index.ts
@@ -1,2 +1,2 @@
-export { SectionsList } from "./lib/cuicui-components/sections-list";
+export { SectionsList } from "./lib/sections-list";
 export { AnimatedNumber } from "@/cuicui/other/transition-wrappers/animated-number/animated-number";
